Move webscraper to new folder
