+++ back/scraper/.gitignore
@@ -1,0 +1,2 @@
+node_modules
+package-lock.json
