+++ back/scraper/keywords.txt
@@ -1,0 +1,6 @@
+title
+#floorplans-1
+.fp-col-text
+.image-link
+img
+src
