+++ back/scraper/package.json
@@ -1,0 +1,7 @@
+{
+  "dependencies": {
+    "axios": "^1.1.3",
+    "cheerio": "^1.0.0-rc.12",
+    "robots-parser": "^3.0.0"
+  }
+}
