+++ back/server/package.json
@@ -1,7 +1,22 @@
 {
+  "name": "housing-prof-server",
+  "version": "1.0.0",
+  "description": "",
+  "main": "server.ts",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "dev": "nodemon server.ts"
+  },
+  "author": "",
+  "license": "ISC",
   "dependencies": {
-    "axios": "^1.1.3",
-    "cheerio": "^1.0.0-rc.12",
-    "robots-parser": "^3.0.0"
+    "express": "^4.18.1"
+  },
+  "devDependencies": {
+    "@types/express": "^4.17.14",
+    "@types/node": "^18.8.2",
+    "nodemon": "^2.0.20",
+    "ts-node": "^10.9.1",
+    "typescript": "^4.8.4"
   }
 }
